+++ components/zcash_address/src/kind/unified/ivk.rs
@@ -1,14 +1,11 @@
 use alloc::vec::Vec;
-use core::{
-    convert::{TryFrom, TryInto},
-    fmt,
-};
+use core::{convert::TryInto, fmt};
 use zcash_protocol::address::Revision;
 use zcash_protocol::constants;
 
 use super::{
     private::{SealedContainer, SealedItem},
-    Container, DataTypecode, Encoding, ParseError, Typecode, Uitem,
+    Container, DataTypecode, Encoding, ParseError, Uitem,
 };
 
 /// The set of known IVKs for Unified IVKs.
@@ -65,37 +62,31 @@ impl fmt::Debug for Ivk {
     }
 }
 
-impl TryFrom<(u32, &[u8])> for Ivk {
-    type Error = ParseError;
-
-    fn try_from((typecode, data): (u32, &[u8])) -> Result<Self, Self::Error> {
+impl SealedItem for Ivk {
+    fn parse(typecode: DataTypecode, data: &[u8]) -> Result<Self, ParseError> {
         let data = data.to_vec();
-        match typecode.try_into()? {
-            Typecode::Data(DataTypecode::P2pkh) => data.try_into().map(Ivk::P2pkh),
-            Typecode::Data(DataTypecode::P2sh) => Err(data),
-            Typecode::Data(DataTypecode::Sapling) => data.try_into().map(Ivk::Sapling),
-            Typecode::Data(DataTypecode::Orchard) => data.try_into().map(Ivk::Orchard),
-            Typecode::Data(DataTypecode::Unknown(_)) => Ok(Ivk::Unknown { typecode, data }),
-            Typecode::Metadata(_) => {
-                return Err(ParseError::InvalidEncoding(format!(
-                    "Unexpected metadata typecode {} in data item position",
-                    typecode
-                )))
-            }
+        match typecode {
+            DataTypecode::P2pkh => data.try_into().map(Ivk::P2pkh),
+            DataTypecode::P2sh => Err(data),
+            DataTypecode::Sapling => data.try_into().map(Ivk::Sapling),
+            DataTypecode::Orchard => data.try_into().map(Ivk::Orchard),
+            DataTypecode::Unknown(tc) => Ok(Ivk::Unknown { typecode: tc, data }),
         }
         .map_err(|e| {
-            ParseError::InvalidEncoding(format!("Invalid ivk for typecode {}: {:?}", typecode, e))
+            ParseError::InvalidEncoding(format!(
+                "Invalid ivk for typecode {}: {:?}",
+                u32::from(typecode),
+                e,
+            ))
         })
     }
-}
 
-impl SealedItem for Ivk {
-    fn typecode(&self) -> Typecode {
+    fn typecode(&self) -> DataTypecode {
         match self {
-            Ivk::P2pkh(_) => Typecode::Data(DataTypecode::P2pkh),
-            Ivk::Sapling(_) => Typecode::Data(DataTypecode::Sapling),
-            Ivk::Orchard(_) => Typecode::Data(DataTypecode::Orchard),
-            Ivk::Unknown { typecode, .. } => Typecode::Data(DataTypecode::Unknown(*typecode)),
+            Ivk::P2pkh(_) => DataTypecode::P2pkh,
+            Ivk::Sapling(_) => DataTypecode::Sapling,
+            Ivk::Orchard(_) => DataTypecode::Orchard,
+            Ivk::Unknown { typecode, .. } => DataTypecode::Unknown(*typecode),
         }
     }
 
@@ -192,10 +183,10 @@ mod tests {
         sample::select,
     };
 
-    use super::{Ivk, ParseError, Typecode, Uivk};
+    use super::{Ivk, ParseError, Uivk};
     use crate::kind::unified::{
         private::{SealedContainer, SealedItem},
-        Container, DataTypecode, Encoding, Revision, Uitem,
+        Container, DataTypecode, Encoding, Revision, Typecode, Uitem,
     };
     use zcash_protocol::consensus::NetworkType;